Add started code for parking lot - 1.

+++ parkinglot/src/main/java/com/example/parkinglot/dtos/CreateTicket.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class CreateTicket {
     String vehicleRegNo;
     VehicleType vehicleType;
-    int gateId;
+    Long gateId;
     SlotType slotType;
 }
